+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## Unreleased 
+## Unreleased
 
-docs: Improve code doc on start and endSession (#1098)
+- feat: Expose SentryDebugImageProvider (#1094)
+- docs: Improve code doc on start and endSession (#1098)
 
 ## 7.0.3