+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Unreleased
 
-- feat: Add automatic performance instrumentation for ui lifecycle and http requests. (#1043)
-- ref: Make calls to customSamplingContext nonnull (#1061)
+- feat: Add UI auto instrumentation (#1105)
 
 ## 7.0.0-beta.0