+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: Expose SentryDebugImageProvider (#1094)
 - docs: Improve code doc on start and endSession (#1098)
+- remove libz (#1109)
 
 ## 7.0.3